Set up getSupply
looks like the architecture of repository, apiservice and viewmodel is working; now you can update circulating value by taping the search icon, which makes no sense but I finally got a grasp on the ktor client

+++ app/src/main/java/com/example/getblock/model/Supply.kt
@@ -14,6 +14,6 @@ data class Result (
 
 @Serializable
 data class Value(
-    val circulating: Int,
-    val nonCirculating: Int
+    val circulating: Long,
+    val nonCirculating: Long
 )
